Last commit before file renaming

+++ src/chr/ft_putchar.c
@@ -6,7 +6,7 @@
 /*   By: gguiulfo <gguiulfo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/03/02 17:43:52 by gguiulfo          #+#    #+#             */
-/*   Updated: 2017/08/21 08:47:14 by gguiulfo         ###   ########.fr       */
+/*   Updated: 2017/11/22 11:23:35 by gguiulfo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,5 +19,5 @@
 
 int	ft_putchar(int c)
 {
-	return (write(STDOUT_FILENO, &c, 1));
+	return ((int)write(STDOUT_FILENO, &c, 1));
 }
